really dumb of me
